+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -27,6 +27,7 @@
 - Support to copy a custom model from one Form Recognizer resource to another
 - Authentication using `azure-identity` credentials now supported
   - see the [Azure Identity documentation](https://github.com/Azure/azure-sdk-for-python/blob/master/sdk/identity/azure-identity/README.md) for more information
+- `page_number` attribute has been added to `FormTable`
 
 ## 1.0.0b2 (2020-05-06)